requirements.txt
asgiref==3.7.2
Django==5.0.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
psycopg==3.1.16
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
